Remove redundant namespaces from enums
As per the "Avoid Redundant Names in Fully-Qualified Namespaces"[1]
section in the Zig language reference, remove name of the enum from each
enum variant. This makes our code a bit more terse and readable.

Raylib used this naming convention because its a C library and C has no
namespacing so it is conventional to include the name of the enum type
in each variant to give it a "namespace". Zig has proper namespacing so
we don't need this.

The only meaningful enum variant change here is
`TraceLogLevel.log_error` which has been renamed to `TraceLogLevel.err`
so as not to clash with Zig's `error` keyword. All other enum variants
simply have their enum name removed.

This commit also fixes up example code, and switches out the full enum
paths to just use an enum literal, as the context is already clear.

[1]https://ziglang.org/documentation/master/#Avoid-Redundant-Names-in-Fully-Qualified-Namespaces